Use cmake+ninja by default (#4849)
* cmake+ninja: Allow using ninja with cmake

* libaom: Disable tests to expedite compile time

* cmake.mk: Set default TOOLCHAIN location and prefix

Instead of setting each tool as CMAKE options, use
the default variables to set the default path and prefix.

* cmake.mk: Set CMAKE_ASM_COMPILER to nasm as needed

Also, if nasm is not set, we shouldn't unset ENABLE_ASSEMBLY.

* cmake+ninja: Allow using CMAKE_USE_DESTDIR from ninja

The CMAKE_USE_DESTDIR is by default enabled (i.e. set to 1)
and doing so adds to the environment DESTDIR=<install_path>
to the installation call. Some dependencies will not install
properly when DESTDIR is set (i.e. umurmur and intel-media-driver).
As such CMAKE_USE_DESTDIR must be set to 0 so the DESTDIR variable
is not being set in the environment when installing.

This functionality is now added to ninja to allow not setting
DESTDIR when needed.

* ninja.mk: Convert CMAKE_USE_DESTDIR to NINJA_USE_DESTDIR

This allows having this functionality for other use cases such as
meson+ninja where we could define NINJA_USE_DESTDIR to 0 or 1 as
needed.